Conflicts:
	Your_Location_In_The_Nation/.classpath
	Your_Location_In_The_Nation/bin/.gitignore
	Your_Location_In_The_Nation/bin/DatabasePersist/DerbyDatabase$2.class
	Your_Location_In_The_Nation/derby.log

+++ Your_Location_In_The_Nation/src/FakeDatabase/InitialData.java
@@ -196,7 +196,7 @@ public static List<CostOfLiving> getCostOfLivingRent() throws IOException {
 			Iterator<String> i = tuple.iterator();
 			CostOfLiving COL = new CostOfLiving();
 			COL.setScale(Integer.parseInt(i.next()));
-			COL.setCostOfLivingIndex(Integer.parseInt(i.next()));
+			COL.setCostOfLivingIndex(Float.parseFloat(i.next()));
 			
 			
 			CostOfLivingList.add(COL);
@@ -223,7 +223,7 @@ public static List<CostOfLiving> getCostOfLivingMortgage() throws IOException {
 			Iterator<String> i = tuple.iterator();
 			CostOfLiving COL = new CostOfLiving();
 			COL.setScale(Integer.parseInt(i.next()));
-			COL.setCostOfLivingIndex(Integer.parseInt(i.next()));
+			COL.setCostOfLivingIndex(Float.parseFloat(i.next()));
 			
 			
 			CostOfLivingList.add(COL);
@@ -250,7 +250,7 @@ public static List<CostOfLiving> getCostOfLivingNoMortgage() throws IOException 
 			Iterator<String> i = tuple.iterator();
 			CostOfLiving COL = new CostOfLiving();
 			COL.setScale(Integer.parseInt(i.next()));
-			COL.setCostOfLivingIndex(Integer.parseInt(i.next()));
+			COL.setCostOfLivingIndex(Float.parseFloat(i.next()));
 			
 			
 			CostOfLivingList.add(COL);
